Rework package structure to make more sense of it
Follow a similar approach to the modules RFC in Ember.
The default export (if any) is a class. Types and functions are named exports.

Removed the single entry point. Now there is two distinct entry points:

`ember-ast-helpers/built-time-component` for BuildTimeComponent and related helpers
`ember-ast-helpers/html` for html-related helpers not specifically tied to BuildTimeComponent